requirements.txt
beautifulsoup4
faiss-cpu
openai
tiktoken
tqdm
numpy
python-dotenv